Begin to write a own template engine
This is not really necessary, because it exist enough good well tested template
engines. But I want to learn java with this project. Therefore I'll use my own
template engine for this task.

The first version from this engine can only replace strings with an hashmap.
This syntax is inspired by the Jinja template engine.